Adding comments where they were missing and reorganizing the project layout
