+++ Solnet.Mango.Examples/CancelPerpOrderExample.cs
@@ -1,5 +1,6 @@
 ﻿using Solnet.KeyStore;
 using Solnet.Mango.Models;
+using Solnet.Mango.Models.Matching;
 using Solnet.Mango.Models.Perpetuals;
 using Solnet.Programs;
 using Solnet.Programs.Models;
