+++ Solnet.Mango.Examples/GetOrderBookSideExample.cs
@@ -1,4 +1,5 @@
 using Solnet.Mango.Models;
+using Solnet.Mango.Models.Matching;
 using Solnet.Mango.Models.Perpetuals;
 using Solnet.Programs;
 using Solnet.Programs.Models;
