+++ Solnet.Mango.Test/TestBase.cs
@@ -1,4 +1,5 @@
 ﻿using Solnet.Mango.Models;
+using Solnet.Mango.Models.Banks;
 using Solnet.Programs;
 using Solnet.Serum.Models;
 using System;
