+++ Solnet.Mango/Models/Enums/DataType.cs
