+++ Solnet.Mango/Models/Enums/MarketType.cs
@@ -1,0 +1,18 @@
+namespace Solnet.Mango.Models
+{
+    /// <summary>
+    /// Represents the types of markets available.
+    /// </summary>
+    public enum MarketType : byte
+    {
+        /// <summary>
+        /// A spot margined market.
+        /// </summary>
+        Spot,
+
+        /// <summary>
+        /// A perpetual market.
+        /// </summary>
+        Perpetual
+    }
+}
