--- Solnet.Mango/Models/MarketType.cs
@@ -1,18 +1,0 @@
-namespace Solnet.Mango.Models
-{
-    /// <summary>
-    /// Represents the types of markets available.
-    /// </summary>
-    public enum MarketType : byte
-    {
-        /// <summary>
-        /// A spot margined market.
-        /// </summary>
-        Spot,
-
-        /// <summary>
-        /// A perpetual market.
-        /// </summary>
-        Perpetual
-    }
-}
